// src/MultiSchema/types.ts
export interface ImportInfo {
	importName: string;
	importKind: 'named' | 'default' | 'namespace';
	modulePath: string;
	isTypeOnly?: boolean;
}